// frontend/src/app/core/services/api/model/scoringSchemeDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ScoringType } from './scoringType';


/**
 * Data Transfer Object for scoring scheme
 */
export interface ScoringSchemeDto { 
    uuid?: string;
    name?: string;
    scoringType?: ScoringType;
    isActive?: boolean;
    createdAt?: string;
    modifiedAt?: string;
}
export namespace ScoringSchemeDto {
}