// frontend/src/app/core/services/api/model/scoringSchemeBoatCoefficientRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BoatClass } from './boatClass';


/**
 * Request payload for creating or updating a scoring scheme boat coefficient
 */
export interface ScoringSchemeBoatCoefficientRequest { 
    scoringSchemeId: string;
    boatClass: BoatClass;
    coefficient: number;
}
export namespace ScoringSchemeBoatCoefficientRequest {
}